fix: move cidlist request to _id page and add loader in asyncData hook
