+++ packages/fe/pages/dataset/_id.vue
@@ -197,7 +197,7 @@
 
 <script>
 // ====================================================================== Import
-import { mapGetters } from 'vuex'
+import { mapGetters, mapActions } from 'vuex'
 
 import TextBlock from '@/components/blocks/text-block'
 import CardCutout from '@/components/card-cutout'
@@ -235,6 +235,7 @@ export default {
   },
 
   async asyncData ({ store, route, error }) {
+    store.dispatch('cid/setLoadingStatus', { status: true })
     const datasetExists = await store.dispatch('dataset/getDataset', { route })
     if (!datasetExists) { return error('Dataset could not be found.') }
     return { datasetExists }
@@ -256,7 +257,8 @@ export default {
   computed: {
     ...mapGetters({
       siteContent: 'general/siteContent',
-      dataset: 'dataset/dataset'
+      dataset: 'dataset/dataset',
+      cidList: 'cid/cidList'
     }),
     slug () {
       return this.dataset.slug
@@ -290,7 +292,6 @@ export default {
       return this.dataset.createdAt ? this.$moment(this.dataset.createdAt).format('YYYY') : '-'
     },
     datasetSize () {
-      console.log(this.dataset.data_size, this.dataset.data_size ? this.$formatBytes(this.dataset.data_size) : '-')
       return this.dataset.data_size ? this.$formatBytes(this.dataset.data_size) : '-'
     },
     totalDataOnNetwork () {
@@ -332,14 +333,40 @@ export default {
     }
   },
 
-  mounted () {
+  watch: {
+    '$route' () {
+      this.$nextTick(() => {
+        this.getCidList({ route: this.$route })
+      })
+    },
+    cidList () {
+      this.stopLoading()
+    }
+  },
+
+  async mounted () {
     handlePageResize(this)
     this.resize = () => { handlePageResize(this) }
     window.addEventListener('resize', this.resize)
+    await this.getCidList({ route: this.$route })
   },
 
   beforeDestroy () {
     if (this.resize) { window.removeEventListener('resize', this.resize) }
+  },
+
+  methods: {
+    ...mapActions({
+      getCidList: 'cid/getCidList',
+      setLoadingStatus: 'cid/setLoadingStatus'
+    }),
+    stopLoading () {
+      this.$nextTick(() => {
+        if (typeof this.cidList !== 'boolean') {
+          this.setLoadingStatus({ status: false })
+        }
+      })
+    }
   }
 }
 </script>
